+++ src/Euterpe/Features/Share/ShareImportDialog.axaml
@@ -49,7 +49,7 @@
 
         <Grid
             ColumnDefinitions="*,Auto"
-            IsVisible="{Binding !IsImporting, Mode=OneWay}">
+            IsVisible="{Binding !ImportCommand.IsRunning, Mode=OneWay}">
             <Button
                 Classes="Primary"
                 Command="{Binding ImportCommand, Mode=OneTime}"
@@ -70,7 +70,7 @@
 
         <!--  Import Progress  -->
         <StackPanel
-            IsVisible="{Binding IsImporting, Mode=OneWay}"
+            IsVisible="{Binding ImportCommand.IsRunning, Mode=OneWay}"
             Spacing="7">
             <ProgressBar
                 Classes="Primary Rounded"
